API to change user password and other minor changes

+++ src/apis/users/dal.ts
@@ -18,10 +18,10 @@ export default class UserDAL {
     }
   }
 
-  // Get all users
+  // Get all owners
   static async getAllOwners(query?: RequestQuery): Promise<IUsersDoc[]> {
     try {
-      const apiFeatures = new APIFeatures(User.find(), query)
+      const apiFeatures = new APIFeatures(User.find({ role: "Owner" }), query)
         .paginate()
         .filter()
         .project()
@@ -34,8 +34,24 @@ export default class UserDAL {
     }
   }
 
+  // Get all users in DB
+  static async getAllUsers(query?: RequestQuery): Promise<IUsersDoc[]> {
+    try {
+      const apiFeatures = new APIFeatures(User.find(), query)
+        .paginate()
+        .project()
+        .sort()
+        .filter();
+
+      const users = await apiFeatures.dbQuery;
+      return users;
+    } catch (error) {
+      throw error;
+    }
+  }
+
   // Get user by id
-  static async getOwnerById(id: string): Promise<IUsersDoc | null> {
+  static async getUserById(id: string): Promise<IUsersDoc | null> {
     try {
       const user = await User.findById(id);
       return user;
@@ -45,9 +61,9 @@ export default class UserDAL {
   }
 
   // Delete all user in DB
-  static async deleteAll() {
+  static async deleteAllOwners() {
     try {
-      await User.deleteMany();
+      await User.deleteMany({ role: "Owner" });
     } catch (error) {
       throw error;
     }
@@ -58,6 +74,48 @@ export default class UserDAL {
     try {
       const deletedUser = await User.findByIdAndDelete(id);
       return deletedUser;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  // Get user by phone number
+  static async getByPhoneNum(phone_num: string): Promise<IUsersDoc | null> {
+    try {
+      const user = await User.findOne({ phone_num });
+      return user;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  // Update profile
+  static async updateProfile(
+    id: string,
+    data: UserRequests.IUpdateProfile
+  ): Promise<IUsersDoc | null> {
+    try {
+      const user = await User.findByIdAndUpdate(id, data, {
+        runValidators: true,
+        new: true,
+      });
+      return user;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  // Change password
+  static async changePswd(
+    password: string,
+    user: IUsersDoc
+  ): Promise<IUsersDoc | null> {
+    try {
+      user.password = password;
+      user.is_default_pswd = false;
+      user.is_credential_changed = true;
+      await user.save();
+      return user;
     } catch (error) {
       throw error;
     }
